Authentication/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
PyJWT==2.8.0
python-decouple==3.8
pydantic[email]
SQLAlchemy
passlib
psycopg2
bcrypt
streamlit
requests==2.31.0